Got Audio working alootttt cleaner and easier.
FInally figured it out. Need to render as many cycles as I can for the times that I can't.

For mobile, try to implement frameskips, and use performance.now() to get how much time we got in a frame budget

Also, need to play with out audio preference numbers. And can also try to find a balance between time remaining and auduio queue index