Started a memory viewer, need to add breakpoints (#255)
relates to #226

Started a memory viewer, and abstracted all of the virtual list type code stuff. Also, found a huge memory issue in relaying messages, and getting an infinite message listening for our smart workers.

This PR is only half done. but merging now because the between-refresh memory leak is bothering me 😠 

<img width="1280" alt="screen shot 2019-02-15 at 9 51 31 pm" src="https://user-images.githubusercontent.com/1448289/52895199-e97c2600-316b-11e9-8258-1932633e5528.png">